Fixes Issue #28 #29

+++ libraries/helpers.rb
@@ -9,6 +9,7 @@ module Terraform
   module Helpers
     # Transform raw output of the checksum list into a Hash[filename, checksum].
     def raw_checksums_to_hash
+      return {} unless File.exist?(checksums_file_path)
       raw_checksums = File.open(checksums_file_path, 'r').read
       Hash[
         raw_checksums.split("\n").map do |s|
@@ -37,11 +38,11 @@ module Terraform
     def import_gpg_key
       return if key_imported?
       keyfile = File.join(Dir.tmpdir, 'hashicorp.asc')
-      GPGME::Key.import(File.open(keyfile))
+      ::GPGME::Key.import(File.open(keyfile))
     end
 
     def key_imported?
-      !GPGME::Key.find(:public, 'security@hashicorp.com', :sign).empty?
+      !::GPGME::Key.find(:public, 'security@hashicorp.com', :sign).empty?
     end
 
     # verify the sha256sum file's signature
@@ -49,13 +50,16 @@ module Terraform
     def sig_verified?
       checksums = File.open(checksums_file_path, 'r')
       io = File.open(sigfile_path, 'rb')
-      crypto = GPGME::Crypto.new
-      signature = GPGME::Data.new(io)
+      crypto = ::GPGME::Crypto.new
+      signature = ::GPGME::Data.new(io)
       crypto.verify(signature, signed_text: checksums) do |sig|
         return sig.valid? &&
                !(sig.expired_signature? || sig.expired_key? ||
                  sig.revoked_key? || sig.bad? || sig.no_key?)
       end
+    rescue Errno::ENOENT
+      false
+    ensure
       false
     end
